requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
tenacity==9.0.0
gunicorn==23.0.0